make all method receivers, pointers, for performance reasons
Initially they were converted to copies (not pointers)
so that the struct can't be changed during runtime by the
wrong methods, but the performance cost isn't worth it.

They can be made immutable by just not exporting them,
but it might be useful to keep them as exported,
because makes interfacing outside the package easier.
Sometimes it might also be the desired behavior,
because outside the package, code might need to
change a value during runtime to avoid restarting
the program or reinitialising the whole struct.

Getters and setters wouldn't be useful in this case.